basic product create

+++ src/main/java/com/example/ecommercebackend/service/file/CategoryImageService.java
@@ -35,7 +35,7 @@ public class CategoryImageService implements ImageService<CoverImage,ImageReques
 
     @Transactional
     @Override
-    public CoverImage save(ImageRequestDto file, Long id) {
+    public CoverImage save(ImageRequestDto file, Integer id) {
 
         long maxSize = productImageSize;
         if (file.getMultipartFile().getSize() > maxSize) {
@@ -67,7 +67,7 @@ public class CategoryImageService implements ImageService<CoverImage,ImageReques
 
     @Transactional
     @Override
-    public String delete(Long id) {
+    public String delete(Integer id) {
         CoverImage coverImage = getById(id);
         String path = coverImage.getUrl().replace(uploadFileUrl,uploadFileDir);
         System.out.println("pppath: "+path);
@@ -77,8 +77,8 @@ public class CategoryImageService implements ImageService<CoverImage,ImageReques
     }
 
     @Override
-    public CoverImage getById(Long id) {
-        return coverImageRepository.findById(id).orElseThrow(() -> new BadRequestException("CoverImage not found."));
+    public CoverImage getById(Integer id) {
+        return coverImageRepository.findById(Long.valueOf(id)).orElseThrow(() -> new BadRequestException("CoverImage not found."));
     }
 
     @Override
