+++ src/main/java/com/example/ecommercebackend/service/file/IFileService.java
@@ -7,9 +7,9 @@ import com.example.ecommercebackend.service.storagestrategy.IStorageStrategy;
 import java.util.List;
 
 public interface IFileService<T extends File,R extends FileRequestDto> {
-    T save(R file,Long id);  // Dosya kaydetme
-    String delete(Long id);  // Dosya silme
-    T getById(Long id);  // Belirli ID'ye sahip dosyayı getir
+    T save(R file,Integer id);  // Dosya kaydetme
+    String delete(Integer id);  // Dosya silme
+    T getById(Integer id);  // Belirli ID'ye sahip dosyayı getir
     List<T> getAll();  // Tüm dosyaları getir
     void setStorageStrategy(IStorageStrategy strategy); // Depolama türünü belirle
     String getFileExtension(String fileName);
